Automated canary cut from `main` (`74b9042347`).

**Merge with "Create a merge commit"** so the two-parent
merge commit survives into `canary`'s history.

+++ contrib/rust-sdk/perfetto/src/protos/trace/track_event/thread_descriptor.pz.rs
@@ -38,7 +38,7 @@ pb_enum!(ThreadDescriptorChromeThreadType {
 
 pb_msg!(ThreadDescriptor {
     pid: i32, primitive, 1,
-    tid: i32, primitive, 2,
+    tid: i64, primitive, 2,
     thread_name: String, primitive, 5,
     chrome_thread_type: ThreadDescriptorChromeThreadType, enum, 4,
     reference_timestamp_us: i64, primitive, 6,
